+++ config/routes.rb
@@ -123,13 +123,14 @@ Rails.application.routes.draw do
     end
   end
 
-  get 'research-topics/accepted/does-treatment-of-sleep-apnea-influence-body-weight', to: "research_topics#sleep_apnea_body_weight", as: "sleep_apnea_body_weight"
-  get 'research-topics/accepted/does-sleep-influence-memory-and-brain-plasticity', to: "research_topics#sleep_apnea_brain_plasticity", as: "sleep_apnea_brain_plasticity"
-  get 'research-topics/accepted/obstructive-sleep-apnea-and-adenotonsillectomy-in-children', to: "research_topics#sleep_apnea_adenotonsillectomy_children", as: "sleep_apnea_adenotonsillectomy_children"
-  get 'research-topics/accepted/link-between-type-2-diabetes-and-sleep-apnea', to: "research_topics#sleep_apnea_diabetes", as: "sleep_apnea_diabetes"
-  get 'research-topics/accepted/can-nighttime-oxygen-replace-cpap-for-treatment-of-sleep-apnea', to: "research_topics#sleep_apnea_nighttime_oxygen_use", as: "sleep_apnea_nighttime_oxygen_use"
-  get 'research-topics/accepted/didgeridoo-a-potentially-novel-intervention-for-sleep-apnea', to: "research_topics#sleep_apnea_didgeridoo", as: "sleep_apnea_didgeridoo"
-  get 'research-topics/accepted/unilateral-hypoglossal-nerve-stimulation-sleep-apnea-treatment', to: "research_topics#sleep_apnea_hypoglossal_stimulation", as: "sleep_apnea_hypoglossal_stimulation"
+  get 'research-topics/accepted/does-treatment-of-sleep-apnea-influence-body-weight', to: 'research_topics#sleep_apnea_body_weight', as: 'sleep_apnea_body_weight'
+  get 'research-topics/accepted/does-sleep-influence-memory-and-brain-plasticity', to: 'research_topics#sleep_apnea_brain_plasticity', as: 'sleep_apnea_brain_plasticity'
+  get 'research-topics/accepted/obstructive-sleep-apnea-and-adenotonsillectomy-in-children', to: 'research_topics#sleep_apnea_adenotonsillectomy_children', as: 'sleep_apnea_adenotonsillectomy_children'
+  get 'research-topics/accepted/link-between-type-2-diabetes-and-sleep-apnea', to: 'research_topics#sleep_apnea_diabetes', as: 'sleep_apnea_diabetes'
+  get 'research-topics/accepted/can-nighttime-oxygen-replace-cpap-for-treatment-of-sleep-apnea', to: 'research_topics#sleep_apnea_nighttime_oxygen_use', as: 'sleep_apnea_nighttime_oxygen_use'
+  get 'research-topics/accepted/didgeridoo-a-potentially-novel-intervention-for-sleep-apnea', to: 'research_topics#sleep_apnea_didgeridoo', as: 'sleep_apnea_didgeridoo'
+  get 'research-topics/accepted/unilateral-hypoglossal-nerve-stimulation-sleep-apnea-treatment', to: 'research_topics#sleep_apnea_hypoglossal_stimulation', as: 'sleep_apnea_hypoglossal_stimulation'
+  get 'research-topics/accepted/women-with-sleep-apnea-increased-risk-for-heart-disease-with-age', to: 'research_topics#sleep_apnea_women_heart_disease', as: 'sleep_apnea_women_heart_disease'
 
   # Surveys
   resources :surveys do